linuxManualConfig: fix inaccurate FIXME comment
I read this comment, checked out 4.19, saw that it didn't have any
instances of /bin/pwd, and rejoiced, assuming that the removal must
have been backported to the affected kernels.  But actually, /bin/pwd
was just removed earlier, in 4.15, so kernels older than that are
still affected.

Fix the version number in the comment so I don't go through exactly
the same process again a few months from now next time I'm working on
this file.